part-4/src_2/requirements.txt
functions-framework==3.*
google-cloud
google-cloud-secret-manager
google-cloud-storage
PyYAML
pytz
google-cloud-logging
pandas
pandas-gbq
google-cloud-pubsub
google-cloud-firestore
beautifulsoup4